use double instead of float for higher resolution on Mandelbrot value calculations; linearized view snakes around instead of starting from the left at each next row; cursors on both views show the notes being played

+++ src/com/generalprocessingunit/processing/Hilbert.java
@@ -33,7 +33,7 @@ public class Hilbert
             }
 
             //Swap x and y
-            float temp = v.x;
+            double temp = v.x;
             v.x = v.y;
             v.y = temp;
         }
